create useWindowDimensions hook to get screen size in sections
