+++ src/constants/sharedTypes.ts
@@ -1,0 +1,7 @@
+export type SectionProps = {
+  height?: number;
+};
+
+export type SectionContentProps = {
+  width: number;
+};
