+++ src/sections/experience.tsx
@@ -1,16 +1,10 @@
 import * as React from 'react';
 import styled from 'styled-components';
 import { Section, SectionContent } from '../components/shared/sections';
+import { useWindowDimensions } from '../utils/hooks/useWindowDimensions';
 
 const Experience = () => {
-  // TODO: use context and provider to pass window height and width to all sections
-  const [windowHeight, setWindowHeight] = React.useState(0);
-  const [windowWidth, setWindowWidth] = React.useState(0);
-
-  React.useEffect(() => {
-    setWindowHeight(window.innerHeight);
-    setWindowWidth(window.innerWidth);
-  }, []);
+  const { windowWidth, windowHeight } = useWindowDimensions();
 
   return (
     <Section id="experience" height={windowHeight}>
