+++ src/utils/hooks/useWindowDimensions.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export function useWindowDimensions() {
+  const [windowWidth, setWindowWidth] = React.useState(window.innerWidth);
+  const [windowHeight, setWindowHeight] = React.useState(window.innerHeight);
+
+  React.useEffect(() => {
+    function handleResize() {
+      setWindowWidth(window.innerWidth);
+      setWindowHeight(window.innerHeight);
+    }
+
+    window.addEventListener('resize', handleResize);
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return { windowWidth, windowHeight };
+}
